+++ examples/storybook/stories/Plugins/Image/ImagePluginStory.js
@@ -12,25 +12,11 @@ import editorSourcecode from '!!raw-loader!./ImageEditor.js';
 import ImageViewer from './ImageViewer';
 import viewerSourcecode from '!!raw-loader!./ImageViewer.js';
 import SyntaxHighlighter from '../../Components/SyntaxHighlighter';
-
-const mockData = {
-  id: '8b72558253b2502b401bb46e5599f22a',
-  original_file_name: '8bb438_1b73a6b067b24175bd087e86613bd00c.jpg', //eslint-disable-line
-  file_name: '8bb438_1b73a6b067b24175bd087e86613bd00c.jpg', //eslint-disable-line
-  width: 1920,
-  height: 1000,
-};
+import { mockImageNativeUploadFunc } from '../../../../main/shared/utils/fileUploadUtil';
 
 const mockErrorMsg = 'file too large';
 const onFilesChangeMap = {
-  mock: (files, updateEntity) => {
-    setTimeout(() => {
-      updateEntity({
-        data: mockData,
-        files,
-      });
-    }, 2000);
-  },
+  mock: mockImageNativeUploadFunc,
   error: (files, updateEntity) => {
     setTimeout(() => {
       updateEntity({ error: { msg: mockErrorMsg } });