+++ examples/storybook/stories/Ricos/index.js
@@ -4,10 +4,12 @@ import ExampleApp from './ExampleAppStory';
 import ThemesStory from './Themes';
 import RicosStaticToolbar from './RicosStaticToolbar';
 import RicosModalSettings from './RicosModalSettings';
+import RicosMediaStory from './RicosMediaStory';
 
 storiesOf('Ricos', module)
   .add('Basic Usage', BasicUsageStory)
   .add('Example App', ExampleApp)
   .add('Palettes', ThemesStory)
   .add('Static Toolbar', RicosStaticToolbar)
-  .add('Modal API', RicosModalSettings);
+  .add('Modal API', RicosModalSettings)
+  .add('Ricos Media', RicosMediaStory);