+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -13,6 +13,7 @@ import {
   PreviewSettings,
   CreatePluginFunction,
 } from './types';
+import { DRAFT_EDITOR_PROPS } from './consts';
 
 export interface RichContentProps {
   config?: Record<string, unknown>;
@@ -65,6 +66,7 @@ export interface RicosEditorProps extends RicosProps {
   onChange?: OnContentChangeFunction;
   placeholder?: string;
   toolbarSettings?: ToolbarSettings;
+  onBusyChange?: OnBusyChangeFunction;
 }
 
 export interface RicosViewerProps extends RicosProps {
@@ -103,19 +105,10 @@ export type OnContentChangeFunction = (content: RicosContent) => void;
 
 export type OnErrorFunction = (error: string) => void;
 
+export type OnBusyChangeFunction = (isBusy: boolean) => void;
+
 // draft-js props - https://draftjs.org/docs/api-reference-editor
-export type DraftEditorSettings = Pick<
-  EditorProps,
-  | 'autoCapitalize'
-  | 'autoComplete'
-  | 'autoCorrect'
-  | 'spellCheck'
-  | 'stripPastedStyles'
-  | 'handleBeforeInput'
-  | 'handlePastedText'
-  | 'handleReturn'
-  | 'tabIndex'
->;
+export type DraftEditorSettings = Pick<EditorProps, typeof DRAFT_EDITOR_PROPS[number]>;
 
 export interface MediaSettings {
   pauseMedia?: boolean;