+++ packages/ricos-common/web/src/consts.ts
@@ -1,0 +1,11 @@
+export const DRAFT_EDITOR_PROPS = [
+  'autoCapitalize',
+  'autoComplete',
+  'autoCorrect',
+  'spellCheck',
+  'stripPastedStyles',
+  'handleBeforeInput',
+  'handlePastedText',
+  'handleReturn',
+  'tabIndex',
+] as const;