+++ packages/ricos-common/web/src/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export { RicosEngine } from './RicosEngine';
 export { shouldRenderChild, emptyState } from './utils';
 export { default as localeStrategy } from './localeStrategy/localeStrategy';
+export { DRAFT_EDITOR_PROPS } from './consts';